requirements.txt
Django~=5.1.1
djangorestframework~=3.15.2
requests~=2.32.3
Faker~=30.3.0